Add serialization utilities for qmodels (#39)
Fixes the GenerateConfig serialization and adds utilities and tests to
verify quantized models can be store and retrieved.

The only supported save format is `'tf'`, since `.keras` is not
supported by `tensorflow-model-optimization` and it's not working with
`.h5` because of a weird hierarchy, which makes the load of the weights
compute incorrectly.